tests/wpt/tests/css/css-size-adjust/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Mobile Text Size Adjustment properties</title>
<link rel="help" href="https://drafts.csswg.org/css-size-adjust/#property-index">
<meta name="assert" content="text-size-adjust inherits according to the spec.">
<meta name="assert" content="text-size-adjust has initial value according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_inherited('text-size-adjust', 'auto', '10%');
</script>
</body>
</html>
